Use double instead of int for Session weights.
Interestingly enough it did not occur to me during the design, but it is
quite common to use weights with decimal values. A lot of weight
machines for example have half weights that can be added. To make this
possible we now use double for all weight values in the app.

Note that it is not yet actually possible to add a session with a
decimal value weight. We will have to update the NumberPicker for this.

+++ app/src/main/java/com/dermobbbda/gymcoachbob/Exercise.java
@@ -18,7 +18,7 @@ public class Exercise implements Serializable {
     /* The default number of Sessions to allocate for an Exercise. */
     private static final int DEFAULT_CAPACITY = 0;
     /* The weight used for the last Session that was added. */
-    private int mLastWeight = 0;
+    private double mLastWeight = 0;
     /* The number of repetitions used for the last Session that was added. */
     private int mLastRepetitions = 0;
 
@@ -70,7 +70,7 @@ public class Exercise implements Serializable {
     }
 
     /* Return the weight that was used for the last Session. */
-    public int lastWeight() {
+    public double lastWeight() {
         return mLastWeight;
     }
 
@@ -87,9 +87,9 @@ class Session implements Serializable, Comparable<Session> {
     /** Number of repetitions of the set. */
     private int mRepetitions;
     /** Weight used for the set. */
-    private int mWeight;
+    private double mWeight;
 
-    Session(Date date, int weight, int repetitions) {
+    Session(Date date, double weight, int repetitions) {
         mDate = date;
         mWeight = weight;
         mRepetitions = repetitions;
@@ -101,7 +101,7 @@ class Session implements Serializable, Comparable<Session> {
     }
 
     /** Return the weight used for this Session. */
-    public int weight() {
+    public double weight() {
         return mWeight;
     }
